# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/portfolios/public": {
      "get": {
        "description": "Returns a published portfolio by slug and name. Increments view count. No auth required.",
        "operationId": "PortfoliosController_getPublicPortfolio",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "query",
            "description": "URL-friendly portfolio identifier",
            "schema": {
              "example": "my-portfolio",
              "type": "string"
            }
          },
          {
            "name": "name",
            "required": true,
            "in": "query",
            "description": "Portfolio name (unique with slug)",
            "schema": {
              "example": "My Portfolio",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Portfolio found and returned",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 200,
                    "message": "Operation successfully!",
                    "data": {
                      "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                      "name": "My Portfolio",
                      "slug": "my-portfolio",
                      "published": true,
                      "status": "ACTIVE",
                      "theme": {
                        "primary": "#000000",
                        "secondary": "#FFFFFF"
                      },
                      "sections": [
                        {
                          "type": "hero",
                          "title": "My Portfolio",
                          "description": "Welcome to my work"
                        }
                      ],
                      "metaTitle": "My Portfolio",
                      "metaDescription": "A showcase of my projects",
                      "favicon": null,
                      "views": 42,
                      "userId": "user-uuid-here",
                      "createdAt": "2025-02-14T10:00:00.000Z",
                      "updatedAt": "2025-02-14T12:00:00.000Z",
                      "deletedAt": null
                    },
                    "timestamp": "2025-02-14T12:00:00.000Z"
                  }
                }
              }
            }
          }
        },
        "summary": "Get public portfolio by slug and name",
        "tags": [
          "portfolios"
        ]
      }
    },
    "/portfolios": {
      "post": {
        "description": "Creates a portfolio for the authenticated user.",
        "operationId": "PortfoliosController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePortfolioDto"
              },
              "examples": {
                "minimal": {
                  "summary": "Minimal (name + slug only)",
                  "value": {
                    "name": "My Portfolio",
                    "slug": "my-portfolio"
                  }
                },
                "full": {
                  "summary": "Full payload",
                  "value": {
                    "name": "My Portfolio",
                    "slug": "my-portfolio",
                    "theme": {
                      "primary": "#000000",
                      "secondary": "#FFFFFF"
                    },
                    "sections": [
                      {
                        "type": "hero",
                        "title": "My Portfolio",
                        "description": "My Portfolio Description"
                      }
                    ],
                    "published": false,
                    "metaTitle": "My Portfolio",
                    "metaDescription": "My Portfolio Description"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Portfolio created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 200,
                    "message": "Operation successfully!",
                    "data": {
                      "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                      "name": "My Portfolio",
                      "slug": "my-portfolio",
                      "published": true,
                      "status": "ACTIVE",
                      "theme": {
                        "primary": "#000000",
                        "secondary": "#FFFFFF"
                      },
                      "sections": [
                        {
                          "type": "hero",
                          "title": "My Portfolio",
                          "description": "Welcome to my work"
                        }
                      ],
                      "metaTitle": "My Portfolio",
                      "metaDescription": "A showcase of my projects",
                      "favicon": null,
                      "views": 42,
                      "userId": "user-uuid-here",
                      "createdAt": "2025-02-14T10:00:00.000Z",
                      "updatedAt": "2025-02-14T12:00:00.000Z",
                      "deletedAt": null
                    },
                    "timestamp": "2025-02-14T12:00:00.000Z"
                  }
                }
              }
            }
          }
        },
        "summary": "Create a new portfolio",
        "tags": [
          "portfolios"
        ]
      },
      "get": {
        "description": "Returns all portfolios for the authenticated user.",
        "operationId": "PortfoliosController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of portfolios",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 200,
                    "message": "Operation successfully!",
                    "data": [
                      {
                        "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                        "name": "My Portfolio",
                        "slug": "my-portfolio",
                        "published": true,
                        "status": "ACTIVE",
                        "theme": {
                          "primary": "#000000",
                          "secondary": "#FFFFFF"
                        },
                        "sections": [
                          {
                            "type": "hero",
                            "title": "My Portfolio",
                            "description": "Welcome to my work"
                          }
                        ],
                        "metaTitle": "My Portfolio",
                        "metaDescription": "A showcase of my projects",
                        "favicon": null,
                        "views": 42,
                        "userId": "user-uuid-here",
                        "createdAt": "2025-02-14T10:00:00.000Z",
                        "updatedAt": "2025-02-14T12:00:00.000Z",
                        "deletedAt": null
                      }
                    ],
                    "timestamp": "2025-02-14T12:00:00.000Z"
                  }
                }
              }
            }
          }
        },
        "summary": "List my portfolios",
        "tags": [
          "portfolios"
        ]
      }
    },
    "/portfolios/{id}": {
      "get": {
        "description": "Returns a single portfolio by ID if it belongs to the current user.",
        "operationId": "PortfoliosController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Portfolio UUID",
            "schema": {
              "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Portfolio found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 200,
                    "message": "Operation successfully!",
                    "data": {
                      "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                      "name": "My Portfolio",
                      "slug": "my-portfolio",
                      "published": true,
                      "status": "ACTIVE",
                      "theme": {
                        "primary": "#000000",
                        "secondary": "#FFFFFF"
                      },
                      "sections": [
                        {
                          "type": "hero",
                          "title": "My Portfolio",
                          "description": "Welcome to my work"
                        }
                      ],
                      "metaTitle": "My Portfolio",
                      "metaDescription": "A showcase of my projects",
                      "favicon": null,
                      "views": 42,
                      "userId": "user-uuid-here",
                      "createdAt": "2025-02-14T10:00:00.000Z",
                      "updatedAt": "2025-02-14T12:00:00.000Z",
                      "deletedAt": null
                    },
                    "timestamp": "2025-02-14T12:00:00.000Z",
                    "user": {
                      "id": "user-uuid-here",
                      "name": "Jane Doe",
                      "email": "jane@example.com"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Portfolio not found"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Get portfolio by ID",
        "tags": [
          "portfolios"
        ]
      },
      "patch": {
        "description": "Updates a portfolio. Only provided fields are updated.",
        "operationId": "PortfoliosController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Portfolio UUID",
            "schema": {
              "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePortfolioDto"
              },
              "examples": {
                "updateName": {
                  "summary": "Update name only",
                  "value": {
                    "name": "Updated Portfolio Name"
                  }
                },
                "publish": {
                  "summary": "Publish portfolio",
                  "value": {
                    "published": true
                  }
                },
                "themeAndMeta": {
                  "summary": "Update theme and meta",
                  "value": {
                    "theme": {
                      "primary": "#1a1a1a",
                      "secondary": "#f5f5f5"
                    },
                    "metaTitle": "New Meta Title",
                    "metaDescription": "New meta description"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Portfolio updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 200,
                    "message": "Operation successfully!",
                    "data": {
                      "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                      "name": "My Portfolio",
                      "slug": "my-portfolio",
                      "published": true,
                      "status": "ACTIVE",
                      "theme": {
                        "primary": "#000000",
                        "secondary": "#FFFFFF"
                      },
                      "sections": [
                        {
                          "type": "hero",
                          "title": "My Portfolio",
                          "description": "Welcome to my work"
                        }
                      ],
                      "metaTitle": "My Portfolio",
                      "metaDescription": "A showcase of my projects",
                      "favicon": null,
                      "views": 42,
                      "userId": "user-uuid-here",
                      "createdAt": "2025-02-14T10:00:00.000Z",
                      "updatedAt": "2025-02-14T12:00:00.000Z",
                      "deletedAt": null
                    },
                    "timestamp": "2025-02-14T12:00:00.000Z"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Portfolio not found"
          }
        },
        "summary": "Update portfolio",
        "tags": [
          "portfolios"
        ]
      },
      "delete": {
        "description": "Soft-deletes a portfolio (sets deletedAt).",
        "operationId": "PortfoliosController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Portfolio UUID",
            "schema": {
              "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Portfolio soft-deleted successfully (returns updated portfolio with status DELETED)",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 200,
                    "message": "Operation successfully!",
                    "data": {
                      "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                      "name": "My Portfolio",
                      "slug": "my-portfolio",
                      "published": true,
                      "status": "ACTIVE",
                      "theme": {
                        "primary": "#000000",
                        "secondary": "#FFFFFF"
                      },
                      "sections": [
                        {
                          "type": "hero",
                          "title": "My Portfolio",
                          "description": "Welcome to my work"
                        }
                      ],
                      "metaTitle": "My Portfolio",
                      "metaDescription": "A showcase of my projects",
                      "favicon": null,
                      "views": 42,
                      "userId": "user-uuid-here",
                      "createdAt": "2025-02-14T10:00:00.000Z",
                      "updatedAt": "2025-02-14T12:00:00.000Z",
                      "deletedAt": null
                    },
                    "timestamp": "2025-02-14T12:00:00.000Z",
                    "status": "DELETED",
                    "deletedAt": "2025-02-14T12:00:00.000Z"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Portfolio not found"
          }
        },
        "summary": "Delete portfolio",
        "tags": [
          "portfolios"
        ]
      }
    },
    "/portfolios/{id}/publish": {
      "patch": {
        "description": "Flips the published flag (true → false or false → true).",
        "operationId": "PortfoliosController_togglePublish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Portfolio UUID",
            "schema": {
              "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Publish state toggled",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 200,
                    "message": "Operation successfully!",
                    "data": {
                      "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                      "name": "My Portfolio",
                      "slug": "my-portfolio",
                      "published": true,
                      "status": "ACTIVE",
                      "theme": {
                        "primary": "#000000",
                        "secondary": "#FFFFFF"
                      },
                      "sections": [
                        {
                          "type": "hero",
                          "title": "My Portfolio",
                          "description": "Welcome to my work"
                        }
                      ],
                      "metaTitle": "My Portfolio",
                      "metaDescription": "A showcase of my projects",
                      "favicon": null,
                      "views": 42,
                      "userId": "user-uuid-here",
                      "createdAt": "2025-02-14T10:00:00.000Z",
                      "updatedAt": "2025-02-14T12:00:00.000Z",
                      "deletedAt": null
                    },
                    "timestamp": "2025-02-14T12:00:00.000Z",
                    "published": true
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Portfolio not found"
          }
        },
        "summary": "Toggle portfolio publish state",
        "tags": [
          "portfolios"
        ]
      }
    }
  },
  "info": {
    "title": "Starter Kit API Documentation",
    "description": "API documentation for Starter Kit",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "jwt": {
        "scheme": "Bearer",
        "bearerFormat": "Bearer",
        "description": "[just text field] Please enter token in following format: Bearer <JWT>",
        "name": "Authorization",
        "type": "http",
        "in": "Header"
      }
    },
    "schemas": {
...
      "CreatePortfolioDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the portfolio",
            "example": "My Portfolio"
          },
          "slug": {
            "type": "string",
            "description": "The slug of the portfolio",
            "example": "my-portfolio"
          },
          "theme": {
            "type": "object",
            "description": "The theme of the portfolio",
            "example": {
              "primary": "#000000",
              "secondary": "#FFFFFF"
            }
          },
          "sections": {
            "description": "The sections of the portfolio",
            "example": [
              {
                "type": "hero",
                "title": "My Portfolio",
                "description": "My Portfolio Description"
              }
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "published": {
            "type": "boolean",
            "description": "The published status of the portfolio",
            "example": true
          },
          "metaTitle": {
            "type": "string",
            "description": "The meta title of the portfolio",
            "example": "My Portfolio"
          },
          "metaDescription": {
            "type": "string",
            "description": "The meta description of the portfolio",
            "example": "My Portfolio Description"
          }
        },
        "required": [
          "name",
          "slug",
          "theme",
          "sections",
          "published",
          "metaTitle",
          "metaDescription"
        ]
      },
      "UpdatePortfolioDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the portfolio",
            "example": "My Portfolio"
          },
          "slug": {
            "type": "string",
            "description": "The slug of the portfolio",
            "example": "my-portfolio"
          },
          "theme": {
            "type": "object",
            "description": "The theme of the portfolio",
            "example": {
              "primary": "#000000",
              "secondary": "#FFFFFF"
            }
          },
          "sections": {
            "description": "The sections of the portfolio",
            "example": [
              {
                "type": "hero",
                "title": "My Portfolio",
                "description": "My Portfolio Description"
              }
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "published": {
            "type": "boolean",
            "description": "The published status of the portfolio",
            "example": true
          },
          "metaTitle": {
            "type": "string",
            "description": "The meta title of the portfolio",
            "example": "My Portfolio"
          },
          "metaDescription": {
            "type": "string",
            "description": "The meta description of the portfolio",
            "example": "My Portfolio Description"
          }
        }
      }
    }
  },
  "security": [
    {
      "jwt": []
    }
  ]
}